Add raytune algorithms (#85)
* feat: Add more plots to raytune-analysis command

* fix: add install of seaborn to github tests

* feat: hep_tfds support for raytune

* fix: Update parameter name in raytune config

* fix: Timeout error running Ray Tune on HPC systems

* up: update raytune parameters

* fix: cleaner sbatch scripts for raytune

* fix: bug in raytune sbatch script for flatiron

* chore: add check for specification of raytune local_dir

* feat: add plotting script for exported comet.ml json

* chore: Add raytune default_metric to config file

* feat: add more hyperparameters to tune in raytune

* fix: remove extra tab

* feat: time raytune command and log time

* feat: log best parameters after raytune run

* fix: raytune dropout parameter

The raytune dropout parameter now sets both the
combined_graph_layer dropout and the output decoding dropout.
The combined_graph_layer dropout will be set to dropout/2.

* feat: add Ray Tune search algos and trial schedulers

* feat: fix onecycle schedule bug in number of steps

* chore: add specification of num_samples in search_space.py

* change default Ray Tune metric to val_loss

* change to tune.grid_search in search config

* feat: catch and log OOM errors in Ray Tune hypertuning

* fix: shuffle files in CMSDatasetFactory

* update raytune scripts for Flatiron cluster

* refactor: remove CustomCallback from raytune runs

* reduce raytune timeout limit from 24h to 1h

* chore: do not sync raytune trials to driver

Not necessary since using shared storage

* fix: ensure phase steps to be ingeter in onecycle scheduler

* feat: add option to run raytune without a trial scheduler

* chore: remove old raytune config parameters

* fix: add missing config key

* fix: keras.backend.clear_session() causing Ray Tune error

Clearing the session isn't actually necessary so we remove this call.

* feat: add ability to count skipped configurations from log file

* fix: save model.onnx in output_dir

* fix: save model.onnx in output_dir

* fix: load model.onnx from output_dir

* feat: log skipped configurations using logging

* fix: update pfntuple names in test

* feat: add additional parameters to raytune search space

* fix: total steps in lr schedule

* feat: add option to choose "epoch" as decay_steps

If decay_steps="epoch" in expdecay config, the decay_steps
will be set to the number of steps in an epoch.

* fix: make it possible to interleave only physical or only gun datasets

Previously you had to provide lists of both physical and gun datasets
to use during training.

* feat: add option to use CosineDecay LR schedule

* fix: add set -e to test script

* fix: remove unused function

* don't shuffle files in cms dataset factory

+++ mlpf/tfmodel/onecycle_scheduler.py
@@ -56,7 +56,7 @@ class OneCycleScheduler(LearningRateSchedule):
         else:
             final_lr = lr_max / (final_div)
 
-        phase_1_steps = steps * warmup_ratio
+        phase_1_steps = int(steps * warmup_ratio)
         phase_2_steps = steps - phase_1_steps
 
         self.lr_max = lr_max
